Add Zipcode
Zipcode is disabled by default
When in country input the correct country is filled in then zipcode became not disabled.

Zip Code validates on RegExp fetched from the object, based on the country name introduced by the user. So every country is validated with it's own ZIP code.

Additionaly format of the zip i displaying on input. Also fetched from the object